+++ test/decorators-js/steps/fixtures.js
@@ -1,0 +1,6 @@
+import { test as base } from 'playwright-bdd';
+import { TodoPage } from './poms';
+
+export const test = base.extend({
+  todoPage: ({}, use) => use(new TodoPage()),
+});