+++ test/types-export/steps/fixtures.ts
@@ -1,0 +1,9 @@
+import { test as base, createBdd } from 'playwright-bdd';
+
+class MyFixture {}
+
+export const test = base.extend<{ myFixture: MyFixture }>({
+  myFixture: async ({}, use) => await use(new MyFixture()),
+});
+
+export const { Given, When, Then, Before, BeforeAll } = createBdd(test);