+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased]
 
+## [8.4.2] - 2025-11-26
+* feat: support multiple step decorators on a single method ([#340](https://github.com/vitalets/playwright-bdd/issues/340))
+* minor: evaluate Playwright config for better compatibility ([#352](https://github.com/vitalets/playwright-bdd/issues/352))
+* setup trusted publishing
+
+## [8.4.1] - 2025-09-03
+* feat: explicitly export types ([#322](https://github.com/vitalets/playwright-bdd/issues/322), [#330](https://github.com/vitalets/playwright-bdd/issues/330))
+
 ## [8.4.0] - 2025-08-22
 * improve 'Tags From Path' by excluding the `node_modules` dir from the feature's extraction method. ([#323](https://github.com/vitalets/playwright-bdd/issues/323))
 * support Playwright 1.55
@@ -281,7 +289,11 @@ Version v3 has several major updates based on the feedback.
 ## 1.2.0
 * Initial public release
 
-[unreleased]: https://github.com/vitalets/playwright-bdd/compare/v8.4.0...HEAD
+[unreleased]: https://github.com/vitalets/playwright-bdd/compare/8.4.2...HEAD
+[8.4.1]: https://github.com/vitalets/playwright-bdd/compare/v8.4.0...v8.4.1
+[8.4.1-0]: https://github.com/vitalets/playwright-bdd/compare/v8.4.0...v8.4.1-0
 [8.4.0]: https://github.com/vitalets/playwright-bdd/compare/v8.3.1...v8.4.0
 [8.3.1]: https://github.com/vitalets/playwright-bdd/compare/v8.3.0...v8.3.1
 [8.3.0]: https://github.com/vitalets/playwright-bdd/compare/v8.2.1...v8.3.0
+
+[8.4.2]: https://github.com/vitalets/playwright-bdd/releases/tag/8.4.2